pyproject.toml
[tool.poetry]
name = "ai-research-assistant"
version = "0.1.0"
description = ""
authors = ["Teemu Ruokokoski <teemu.ruokokoski@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
langchain = "^0.3.21"
openai = "^1.68.2"
langchain-openai = "^0.3.9"
langchain-community = "^0.3.20"
pymupdf = "^1.25.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
